Windows build fix
svn:sync-xref-src-repo-rev: r78764

+++ src/VBox/Runtime/common/rand/randopenssl.cpp
@@ -1,4 +1,4 @@
-/* $Id: randopenssl.cpp 41909 2012-06-25 13:17:31Z noreply@oracle.com $ */
+/* $Id: randopenssl.cpp 41910 2012-06-25 13:28:46Z noreply@oracle.com $ */
 /** @file
  * IPRT - Random Numbers and Byte Streams, OpenSSL.
  */
@@ -42,9 +42,9 @@
 static DECLCALLBACK(void) rtRandAdvOpensslGetBytes(PRTRANDINT pThis, uint8_t *pb, size_t cb)
 {
     /** @todo proper error handling? */
-    int rc = RAND_bytes(pb, cb);
+    int rc = RAND_bytes(pb, (int)cb);
     if (RT_UNLIKELY(rc != 1))
-        rc = RAND_pseudo_bytes(pb, cb);
+        rc = RAND_pseudo_bytes(pb, (int)cb);
     AssertReleaseMsg(rc == 1, ("RAND_bytes returned %d (error %d)\n", rc, ERR_get_error()));
 }
